reformate view page
