+++ src/components/DropzoneComponent.tsx
@@ -1,0 +1,43 @@
+import { useCallback } from 'react';
+import { Box, Typography } from '@mui/material';
+import { useDropzone } from 'react-dropzone';
+
+import CloudUploadOutlinedIcon from '@mui/icons-material/CloudUploadOutlined';
+import { theme } from '../theme/defaultTheme';
+
+export const DropzoneComponent: React.FC<{
+  onFileSelect: (file: File) => void;
+}> = ({ onFileSelect }) => {
+  const onDrop = useCallback(
+    (acceptedFiles: File[]) => {
+      if (acceptedFiles.length) onFileSelect(acceptedFiles[0]);
+    },
+    [onFileSelect]
+  );
+
+  const { getRootProps, getInputProps, isDragActive } = useDropzone({
+    accept: { 'image/*': [] },
+    multiple: false,
+    onDrop,
+  });
+
+  return (
+    <Box
+      {...getRootProps()}
+      sx={{
+        border: '2px dashed grey',
+        p: 2,
+        textAlign: 'center',
+        borderRadius: 1,
+        cursor: 'pointer',
+        bgcolor: 'white',
+      }}
+    >
+      <input {...getInputProps()} />
+      <Typography sx={{ color: theme.palette.text.secondary }}>
+        {isDragActive ? 'Release to drop the file here' : 'Upload Image'}
+      </Typography>
+      <CloudUploadOutlinedIcon color="secondary" fontSize="large" />
+    </Box>
+  );
+};
