+++ src/components/FormTextField.tsx
@@ -40,24 +40,6 @@ export const FormTextField: React.FC<FormTextFieldProps> = ({
               multiline={multiline}
               rows={rows}
               error={!!errors[name]}
-              // helperText={errors[name]?.message as string}
-              // sx={{ bgcolor: 'white' }}
-              // slotProps={{
-              //   input: {
-              //     sx: {
-              //       color: theme.palette.text.primary,
-              //     },
-              //   },
-              //   inputLabel: {
-              //     sx: {
-              //       color: theme.palette.text.secondary,
-              //       '& .MuiFormLabel-asterisk': { color: 'red' },
-              //       '& .MuiInputBase-input': {
-              //         color: theme.palette.text.primary,
-              //       },
-              //     },
-              //   },
-              // }}
             />
           )}
         />
