+++ src/routes/DefaultRoutes.tsx
@@ -3,7 +3,7 @@ import { lazy } from 'react';
 const HomePage = lazy(() => import('../pages/Home'));
 const LoginPage = lazy(() => import('../pages/Login'));
 const ContactPage = lazy(() => import('../pages/ContactUs'));
-const EditView = lazy(() => import('../pages/EditView'));
+const EditView = lazy(() => import('../pages/View'));
 
 export const publicRoutes = [
   { index: true, element: <HomePage /> },
